Remove/Refactor: Hacky code for sync clients is hard to maintain; models module is too broad of a name; client module classes had unclear naming
Solution: Restructure and remove the sync wrapper; Move models' functions into new query module and existing files; Rename client classes
