+++ src/aleph/sdk/client/__init__.py
@@ -1,12 +1,10 @@
-from .authenticated import AuthenticatedAlephClient, AuthenticatedUserSessionSync
-from .base import BaseAlephClient, BaseAuthenticatedAlephClient
-from .client import AlephClient, UserSessionSync
+from .abstract import AlephClient, AuthenticatedAlephClient
+from .authenticated_http import AuthenticatedAlephHttpClient
+from .http import AlephHttpClient
 
 __all__ = [
-    "BaseAlephClient",
-    "BaseAuthenticatedAlephClient",
     "AlephClient",
     "AuthenticatedAlephClient",
-    "UserSessionSync",
-    "AuthenticatedUserSessionSync",
+    "AlephHttpClient",
+    "AuthenticatedAlephHttpClient",
 ]
