+++ src/aleph/sdk/client/http.py
@@ -1,20 +1,7 @@
-import asyncio
 import json
 import logging
-import queue
-import threading
 from io import BytesIO
-from typing import (
-    Any,
-    AsyncIterable,
-    Awaitable,
-    Callable,
-    Dict,
-    Iterable,
-    List,
-    Optional,
-    Type,
-)
+from typing import Any, AsyncIterable, Dict, Iterable, List, Optional, Type
 
 import aiohttp
 from aleph_message import parse_message
@@ -23,8 +10,8 @@ from pydantic import ValidationError
 
 from ..conf import settings
 from ..exceptions import FileTooLarge, MessageNotFoundError, MultipleMessagesError
-from ..models.message import MessageFilter, MessagesResponse
-from ..models.post import Post, PostFilter, PostsResponse
+from ..query.filters import MessageFilter, PostFilter
+from ..query.responses import MessagesResponse, Post, PostsResponse
 from ..types import GenericMessage
 from ..utils import (
     Writable,
@@ -32,158 +19,12 @@ from ..utils import (
     copy_async_readable_to_buffer,
     get_message_type_value,
 )
-from .base import BaseAlephClient
-from .utils import T, wrap_async
+from .abstract import AlephClient
 
 logger = logging.getLogger(__name__)
 
 
-async def run_async_watcher(
-    *args, output_queue: queue.Queue, api_server: Optional[str], **kwargs
-):
-    async with AlephClient(api_server=api_server) as session:
-        async for message in session.watch_messages(*args, **kwargs):
-            output_queue.put(message)
-
-
-def watcher_thread(output_queue: queue.Queue, api_server: Optional[str], args, kwargs):
-    asyncio.run(
-        run_async_watcher(
-            output_queue=output_queue, api_server=api_server, *args, **kwargs
-        )
-    )
-
-
-class UserSessionSync:
-    """
-    A sync version of `UserSession`, used in sync code.
-
-    This class is returned by the context manager of `UserSession` and is
-    intended as a wrapper around the methods of `UserSession` and not as a public class.
-    The methods are fully typed to enable static type checking, but most (all) methods
-    should look like this (using args and kwargs for brevity, but the functions should
-    be fully typed):
-
-    >>> def func(self, *args, **kwargs):
-    >>>     return self._wrap(self.async_session.func)(*args, **kwargs)
-    """
-
-    def __init__(self, async_session: "AlephClient"):
-        self.async_session = async_session
-
-    def _wrap(self, method: Callable[..., Awaitable[T]], *args, **kwargs):
-        return wrap_async(method)(*args, **kwargs)
-
-    def get_messages(
-        self,
-        page_size: int = 200,
-        page: int = 1,
-        message_filter: Optional[MessageFilter] = None,
-        ignore_invalid_messages: bool = True,
-        invalid_messages_log_level: int = logging.NOTSET,
-    ) -> MessagesResponse:
-        return self._wrap(
-            self.async_session.get_messages,
-            page_size=page_size,
-            page=page,
-            message_filter=message_filter,
-            ignore_invalid_messages=ignore_invalid_messages,
-            invalid_messages_log_level=invalid_messages_log_level,
-        )
-
-    # @async_wrapper
-    def get_message(
-        self,
-        item_hash: str,
-        message_type: Optional[Type[GenericMessage]] = None,
-        channel: Optional[str] = None,
-    ) -> GenericMessage:
-        return self._wrap(
-            self.async_session.get_message,
-            item_hash=item_hash,
-            message_type=message_type,
-            channel=channel,
-        )
-
-    def fetch_aggregate(
-        self,
-        address: str,
-        key: str,
-    ) -> Dict[str, Dict]:
-        return self._wrap(self.async_session.fetch_aggregate, address, key)
-
-    def fetch_aggregates(
-        self,
-        address: str,
-        keys: Optional[Iterable[str]] = None,
-    ) -> Dict[str, Dict]:
-        return self._wrap(self.async_session.fetch_aggregates, address, keys)
-
-    def get_posts(
-        self,
-        page_size: int = 200,
-        page: int = 1,
-        post_filter: Optional[PostFilter] = None,
-    ) -> PostsResponse:
-        return self._wrap(
-            self.async_session.get_posts,
-            page_size=page_size,
-            page=page,
-            post_filter=post_filter,
-        )
-
-    def download_file(self, file_hash: str) -> bytes:
-        return self._wrap(self.async_session.download_file, file_hash=file_hash)
-
-    def download_file_ipfs(self, file_hash: str) -> bytes:
-        return self._wrap(
-            self.async_session.download_file_ipfs,
-            file_hash=file_hash,
-        )
-
-    def download_file_to_buffer(
-        self, file_hash: str, output_buffer: Writable[bytes]
-    ) -> None:
-        return self._wrap(
-            self.async_session.download_file_to_buffer,
-            file_hash=file_hash,
-            output_buffer=output_buffer,
-        )
-
-    def download_file_ipfs_to_buffer(
-        self, file_hash: str, output_buffer: Writable[bytes]
-    ) -> None:
-        return self._wrap(
-            self.async_session.download_file_ipfs_to_buffer,
-            file_hash=file_hash,
-            output_buffer=output_buffer,
-        )
-
-    def watch_messages(
-        self,
-        message_filter: Optional[MessageFilter] = None,
-    ) -> Iterable[AlephMessage]:
-        """
-        Iterate over current and future matching messages synchronously.
-
-        Runs the `watch_messages` asynchronous generator in a thread.
-        """
-        output_queue: queue.Queue[AlephMessage] = queue.Queue()
-        thread = threading.Thread(
-            target=watcher_thread,
-            args=(
-                output_queue,
-                self.async_session.api_server,
-                message_filter,
-                {},
-            ),
-        )
-        thread.start()
-        while True:
-            yield output_queue.get()
-
-
-class AlephClient(BaseAlephClient):
+class AlephHttpClient(AlephClient):
     api_server: str
     http_session: aiohttp.ClientSession
 
@@ -221,18 +62,7 @@ class AlephClient(BaseAlephClient):
             )
         )
 
-    def __enter__(self) -> UserSessionSync:
-        return UserSessionSync(async_session=self)
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        close_fut = self.http_session.close()
-        try:
-            loop = asyncio.get_running_loop()
-            loop.run_until_complete(close_fut)
-        except RuntimeError:
-            asyncio.run(close_fut)
-
-    async def __aenter__(self) -> "AlephClient":
+    async def __aenter__(self) -> "AlephHttpClient":
         return self
 
     async def __aexit__(self, exc_type, exc_val, exc_tb):
