--- src/aleph/sdk/models/common.py
@@ -1,8 +1,0 @@
-from pydantic import BaseModel
-
-
-class PaginationResponse(BaseModel):
-    pagination_page: int
-    pagination_total: int
-    pagination_per_page: int
-    pagination_item: str
