+++ src/aleph/sdk/query/filters.py
@@ -1,17 +1,9 @@
 from datetime import datetime
-from typing import Dict, Iterable, List, Optional, Union
+from typing import Dict, Iterable, Optional, Union
 
-from aleph_message.models import AlephMessage, MessageType
+from aleph_message.models import MessageType
 
-from .common import PaginationResponse
-from .utils import _date_field_to_float, serialize_list
-
-
-class MessagesResponse(PaginationResponse):
-    """Response from an aleph.im node API on the path /api/v0/messages.json"""
-
-    messages: List[AlephMessage]
-    pagination_item = "messages"
+from ..utils import _date_field_to_float, serialize_list
 
 
 class MessageFilter:
@@ -81,6 +73,73 @@ class MessageFilter:
             ),
             "contentTypes": serialize_list(self.content_types),
             "contentKeys": serialize_list(self.content_keys),
+            "refs": serialize_list(self.refs),
+            "addresses": serialize_list(self.addresses),
+            "tags": serialize_list(self.tags),
+            "hashes": serialize_list(self.hashes),
+            "channels": serialize_list(self.channels),
+            "chains": serialize_list(self.chains),
+            "startDate": _date_field_to_float(self.start_date),
+            "endDate": _date_field_to_float(self.end_date),
+        }
+
+        # Ensure all values are strings.
+        result: Dict[str, str] = {}
+
+        # Drop empty values
+        for key, value in partial_result.items():
+            if value:
+                assert isinstance(value, str), f"Value must be a string: `{value}`"
+                result[key] = value
+
+        return result
+
+
+class PostFilter:
+    """
+    A collection of filters that can be applied on post queries.
+
+    """
+
+    types: Optional[Iterable[str]]
+    refs: Optional[Iterable[str]]
+    addresses: Optional[Iterable[str]]
+    tags: Optional[Iterable[str]]
+    hashes: Optional[Iterable[str]]
+    channels: Optional[Iterable[str]]
+    chains: Optional[Iterable[str]]
+    start_date: Optional[Union[datetime, float]]
+    end_date: Optional[Union[datetime, float]]
+
+    def __init__(
+        self,
+        types: Optional[Iterable[str]] = None,
+        refs: Optional[Iterable[str]] = None,
+        addresses: Optional[Iterable[str]] = None,
+        tags: Optional[Iterable[str]] = None,
+        hashes: Optional[Iterable[str]] = None,
+        channels: Optional[Iterable[str]] = None,
+        chains: Optional[Iterable[str]] = None,
+        start_date: Optional[Union[datetime, float]] = None,
+        end_date: Optional[Union[datetime, float]] = None,
+    ):
+        self.types = types
+        self.refs = refs
+        self.addresses = addresses
+        self.tags = tags
+        self.hashes = hashes
+        self.channels = channels
+        self.chains = chains
+        self.start_date = start_date
+        self.end_date = end_date
+
+    def as_http_params(self) -> Dict[str, str]:
+        """Convert the filters into a dict that can be used by an `aiohttp` client
+        as `params` to build the HTTP query string.
+        """
+
+        partial_result = {
+            "types": serialize_list(self.types),
             "refs": serialize_list(self.refs),
             "addresses": serialize_list(self.addresses),
             "tags": serialize_list(self.tags),
