--- tests/unit/test_synchronous_get.py
@@ -1,20 +1,0 @@
-from aleph_message.models import MessagesResponse, MessageType
-
-from aleph.sdk import AlephClient
-from aleph.sdk.conf import settings
-from aleph.sdk.models.message import MessageFilter
-
-
-def test_get_post_messages():
-    with AlephClient(api_server=settings.API_HOST) as session:
-        response: MessagesResponse = session.get_messages(
-            page_size=2,
-            message_filter=MessageFilter(
-                message_types=[MessageType.post],
-            ),
-        )
-
-        messages = response.messages
-        assert len(messages) > 1
-        for message in messages:
-            assert message.type == MessageType.post
